Can convert from standard NK to measurement NK.
Policies from the topozoo are specified in JSON. There are existing
tools to convert the JSON into standard NetKAT (as defined in the main
frenetic repo). This repo uses similar but distinct types to represent
NetKAT. Thus, we need a way to convert from standard NetKAT to
measurement style NetKAT. The translation is tedious but trivial. It may
need to be tweaked a bit to match up with Haoxian's APIs.