Started to reorganize compiler for benchmarks.
I want to keep the old compiler code path that takes in, out, p and t as
arguments and prints out a bunch of stuff. It's useful for debugging and
we'll need it for some end-to-end tets.

I began changing the topo zoo code path to print out fewer things. I'm
going to change things soon to start printing out times in a nice
format.