Compiler now prints timing info.
The compiler now takes in a policy JSON file, a topology DOT file, and a
list of query files. It iterates through each query, compiles it, counts
the number of predicates, and prints out everything in CSV format. Now,
to generate a bunch of benchmarks, we just have to iterate over the
networks and run them on a select bunch of queries. This will give us a
set of csv files that we can combine to a single big CSV to plot,
tabularize, etc.