Updated Felix README.
A bit has changed since I wrote the last README, so I updated the
README. It should now be descriptive enough for anyone to get the
compiler up and running. In the future, I also want to document how to
run the benchmarks and want to describe some of the small helper scripts
we've written.